+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ============
 
 * Your contribution here.
+* [#23](https://github.com/dblock/guard-rack/pull/23) Supports Windows - [@jonyeezs](https://github.com/jonyeezs)
 
 2.1.1 (03/07/2015)
 ============
@@ -60,4 +61,3 @@
 ================
 
 * Initial public release - [@dblock](https://github.com/dblock).
-